requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
requests>=2.31.0
web3>=6.10.0
pyyaml>=6.0
cryptography>=41.0.0
transformers>=4.35.0
torch>=2.0.0
huggingface-hub>=0.17.0
accelerate>=0.24.0
safetensors>=0.4.0